Capping feature, disabled by default with #undef OPTION_CAPPING in featall.h.
The feature can be set by the config statement "capping <# mips>". The
mips calculation will be restricted to CPs only and capping will only
done on CPs, not on specialty engines.

A monitor task will be started that watches the mipsrate. When that
mipsrate is outside a limit a loop value is adjusted. The instruction 
loop in function run_cpu will loop for the value set by this monitor.

The processor will loop. So it is advisable to keep the number of capped
processors as low as possible. In other words a 10 MIPS capped machine
with 8 CPs can become very hot on looping. Restrict this to 1.


git-svn-id: svn://svn.hercules-390.org/hercules/trunk@6127 969364d3-6d58-4e41-9446-b7e21338ea1b