Performance patch with >10% gain.
I did some profiling with the package perf of Ingo Molnar. After
investigation of the most called functions I implemented the following
features:

The features uses the two-byte opcode table. The second byte can be part
of the instruction format and is used to jump to optimized routines. In
this patch to routines with zero x2.

Optimized BC and BRC
Normally a value of 0x80 is shifted by psw.cc and then masked at the
instruction. The optimized instructions can calculate on psw.cc
directly.

Optimized ST, CL, L, LG and STG
These are RX and RXY instructions, but most of the time the X2 value is
zero. The optimized instructions don't have to calculate and use X2.

run_cpu and run_sie.
I did profiling and testing on those 2. The current setting runs the
fastest on core i7. 

Have fun!

Bernard van der Helm


git-svn-id: svn://svn.hercules-390.org/hercules/trunk@7187 969364d3-6d58-4e41-9446-b7e21338ea1b